Significantly refactored main app file into more manageable lib files.
